Fixed: equals id in main file and temp file

+++ app/src/main/java/files/DataFileManager.java
@@ -45,47 +45,54 @@ public class DataFileManager extends FileManager implements FileWorkMap<String, 
         File file = new File(fileName);
         File tempFile = new File(tempFileName);
         try {
-            if (tempFile.canRead()) {
-                if (isMainFile) {
-                    consoleManager.warning("Внимание! На вашем компьютере был обнаружен временный файл с данными");
-                    if (file.createNewFile()) {
-                        String tempData = readFile(tempFileName);
-                        writeFile(fileName, tempData);
-                        consoleManager.successfully("Данные из временного фалйа были перенесены в основной фалй!");
-                        tempFile.delete();
-                    }
-                    else {
-                        consoleManager.output("Хотите ли вы перенести данные коллекции с временного файла в основной? (yes/no): ");
-                        String response = scanner.nextLine();
-                        boolean isTrueResponse = response.equalsIgnoreCase("yes") || response.equalsIgnoreCase("y")
-                                || response.equalsIgnoreCase("no") || response.equalsIgnoreCase("n");
-
-                        while (!isTrueResponse){
-                            consoleManager.output("Хотите ли вы перенести данные коллекции с временного файла в основной? (yes/no): ");
-                            response = scanner.nextLine();
-                            isTrueResponse = response.equalsIgnoreCase("yes") || response.equalsIgnoreCase("y")
-                                    || response.equalsIgnoreCase("no") || response.equalsIgnoreCase("n");
+            if (!isMainFile){
+                if (tempFile.createNewFile()){
+                    createFile();
+                }
+            } else {
+                if (tempFile.canRead()) {
+                    if (isMainFile) {
+                        consoleManager.warning("Внимание! На вашем компьютере был обнаружен временный файл с данными");
+                        if (file.createNewFile()) {
+                            String tempData = readFile(tempFileName);
+                            writeFile(fileName, tempData);
+                            consoleManager.successfully("Данные из временного фалйа были перенесены в основной фалй!");
+                            tempFile.delete();
                         }
+                        else {
+                            consoleManager.output("Хотите ли вы перенести данные коллекции с временного файла в основной? (yes/no): ");
+                            String response = scanner.nextLine();
+                            boolean isTrueResponse = response.equalsIgnoreCase("yes") || response.equalsIgnoreCase("y")
+                                    || response.equalsIgnoreCase("no") || response.equalsIgnoreCase("n");
 
-                        if (response.equalsIgnoreCase("yes") || response.equalsIgnoreCase("y")) {
-                            Map<String, LabWork> tempCollection = readMap(tempFileName,false, false);
-                            Map<String, LabWork> mainCollection = readMap(getFileName(), true, false);
-                            if (tempCollection != null && mainCollection != null) {
-                                mainCollection.putAll(tempCollection);
-                                if (isCollection(mainCollection)) {
-                                    save(mainCollection);
-                                    consoleManager.successfully("Данные из временного фалйа были перенесены в основной фалй!");
+                            while (!isTrueResponse){
+                                consoleManager.output("Хотите ли вы перенести данные коллекции с временного файла в основной? (yes/no): ");
+                                response = scanner.nextLine();
+                                isTrueResponse = response.equalsIgnoreCase("yes") || response.equalsIgnoreCase("y")
+                                        || response.equalsIgnoreCase("no") || response.equalsIgnoreCase("n");
+                            }
+
+                            if (response.equalsIgnoreCase("yes") || response.equalsIgnoreCase("y")) {
+                                Map<String, LabWork> tempCollection = readMap(tempFileName,false, false);
+                                Map<String, LabWork> mainCollection = readMap(getFileName(), true, false);
+                                if (tempCollection != null && mainCollection != null) {
+                                    mainCollection.putAll(tempCollection);
+                                    if (isCollection(mainCollection, false) != null) {
+                                        save(mainCollection);
+                                        consoleManager.successfully("Данные из временного фалйа были перенесены в основной фалй!");
+                                    } else {
+                                        consoleManager.error("Данные во времнном файле были повреждены, перенос данных невозможен");
+                                    }
                                 } else {
-                                    consoleManager.error("Данные во времнном файле были повреждены, перенос данных невозможен");
+                                    consoleManager.error("Данные во времнном или главном файлах были повреждены, перенос данных невозможен");
                                 }
-                            } else {
-                                consoleManager.error("Данные во времнном или главном файлах были повреждены, перенос данных невозможен");
                             }
                         }
+                        tempFile.delete();
                     }
-                    tempFile.delete();
                 }
             }
+
             if (file.createNewFile()) {
                 createFile();
             }
@@ -96,12 +103,12 @@ public class DataFileManager extends FileManager implements FileWorkMap<String, 
 
     }
 
-    public boolean isCollection(Map<String, LabWork> map) {
-        boolean isCollection = true;
+    public Map<String, LabWork> isCollection(Map<String, LabWork> map, boolean isCreateFile) {
+
+        LabWorkDAO labWorkDAO = new LabWorkDAO();
         try {
             LabWorkChecker labWorkChecker = new LabWorkChecker();
             int maxId = 0;
-            LabWorkDAO labWorkDAO = new LabWorkDAO();
 
             List<Integer> listId = new ArrayList<>();
             for (Map.Entry<String, LabWork> entry : map.entrySet()) {
@@ -121,8 +128,21 @@ public class DataFileManager extends FileManager implements FileWorkMap<String, 
                 Long authorWeight = labWorkChecker.checkWeightPerson(entry.getValue().getAuthor().getWeight().toString(), consoleManager, false);
                 String authorPassportId = labWorkChecker.checkPassportIdPerson(entry.getValue().getAuthor().getPassportID(), consoleManager, false);
 
-                if (listId.contains(id)) {
-                    throw new IOException();
+
+                if (isCreateFile){
+                    if (listId.contains(id)) {
+                        throw new IOException();
+                    }
+                } else {
+                    if (listId.contains(id)) {
+                        id = null;
+                    }
+                    while (id == null){
+                        id = labWorkChecker.checkId(GenerationID.newId().toString(), consoleManager, false);
+                        if (listId.contains(id)) {
+                            id = null;
+                        }
+                    }
                 }
 
                 if (key == null || id == null || dateTime == null || name == null || coordX == null || coordY == null
@@ -133,16 +153,16 @@ public class DataFileManager extends FileManager implements FileWorkMap<String, 
                 listId.add(id);
                 LabWork labWork = new LabWork(id, name, new Coordinates(coordX, coordY), dateTime, minimalPoint, description, difficulty,
                         new Person(authorName, authorWeight, authorPassportId));
-                labWorkDAO.create(key, labWork);
+                map.put(key, labWork);
             }
 
             GenerationID.setId(maxId + 1);
 
         } catch (IOException e) {
-            isCollection = false;
+            return null;
         }
 
-        return isCollection;
+        return map;
     }
 
     @Override
@@ -167,7 +187,7 @@ public class DataFileManager extends FileManager implements FileWorkMap<String, 
 
             labWorkMap = parserJSON.deserializeMap(s);
 
-            if (!isCollection(labWorkMap)) {
+            if (isCollection(labWorkMap, isCreateFile) == null) {
                 throw new IOException();
             }
 
@@ -181,8 +201,8 @@ public class DataFileManager extends FileManager implements FileWorkMap<String, 
             }
 
         } catch (IOException | NullPointerException | DateTimeException e) {
-            consoleManager.error("Во время работы произошла ошибка");
             if (isCreateFile) {
+                consoleManager.error("Во время работы произошла ошибка");
                 createFile();
                 labWorkMap = readMap(fileName, isCreateFile, withMessage);
             }
@@ -223,10 +243,12 @@ public class DataFileManager extends FileManager implements FileWorkMap<String, 
         try {
             if (isMainFile) {
                 consoleManager.warning("Идёт создание файла...");
+                writer = new BufferedWriter(new FileWriter(getFileName()));
             } else {
                 consoleManager.warning("Идёт создание временного файла...");
+                writer = new BufferedWriter(new FileWriter(tempFileName));
             }
-            writer = new BufferedWriter(new FileWriter(getFileName()));
+
             consoleManager.warning("Идёт запись значениями по умолчанию...");
             Map<String, LabWork> labWorkMap = new LinkedHashMap<>();
             LabWork labWork = new LabWork();
